requirements.txt
Flask
google-cloud-storage
pytest
pytest-cov
werkzeug
